requirement.txt
gunicorn==19.9.0
numpy>=1.9.2
scikit-learn>=0.18
pandas>=0.19
streamlit == 0.48.1
lime >= 7.9.0
LightGBM >= 3.2.1
Pillow==2.2.2

